+++ src/hooks/pan/pan-types.ts
@@ -1,0 +1,12 @@
+export interface Coords {
+  x: number
+  y: number
+}
+
+export type PanEvent = {
+  location: Coords
+  delta: Coords
+
+  isFirst: boolean
+  isFinal: boolean
+}
